Update testing functions
Created subdirectory and moved test functions
Changes to be committed:
	renamed:    tests/test_views/__init__.py -> tests/test_apis/__init__.py
	new file:   tests/test_apis/test_api_authentication/__init__.py
	renamed:    tests/test_views/test_account_activate.py -> tests/test_apis/test_api_authentication/test_account_activate.py
	renamed:    tests/test_views/test_account_describe.py -> tests/test_apis/test_api_authentication/test_account_describe.py
	renamed:    tests/test_views/test_account_new.py -> tests/test_apis/test_api_authentication/test_account_new.py
	renamed:    tests/test_views/test_api_auth_add.py -> tests/test_apis/test_api_authentication/test_api_auth_add.py
	renamed:    tests/test_views/test_api_auth_remove.py -> tests/test_apis/test_api_authentication/test_api_auth_remove.py
	renamed:    tests/test_views/test_api_auth_reset_token.py -> tests/test_apis/test_api_authentication/test_api_auth_reset_token.py
	renamed:    tests/test_views/test_api_objects_drafts_create.py -> tests/test_apis/test_api_objects_drafts_create.py
	renamed:    tests/test_views/test_auth_add.py -> tests/test_apis/test_auth_add.py
	renamed:    tests/test_views/test_auth_remove.py -> tests/test_apis/test_auth_remove.py
	renamed:    tests/test_views/test_auth_reset_token.py -> tests/test_apis/test_auth_reset_token.py
	renamed:    tests/test_views/test_objects_drafts_create.py -> tests/test_apis/test_objects_drafts_create.py
	renamed:    tests/test_views/test_prefixes_create.py -> tests/test_apis/test_prefixes_create.py
	new file:   tests/test_apis/test_prefixes_modify.py
	deleted:    tests/test_views/test_api_account_activate.py
	deleted:    tests/test_views/test_api_account_describe.py
	deleted:    tests/test_views/test_api_account_new.py
	deleted:    tests/test_views/test_prefixes_modify.py
