+++ tests/test_apis/test_auth_add.py
