+++ tests/test_apis/test_auth_remove.py
