+++ tests/test_apis/test_prefixes_create.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 
 """Bulk Create Prefixes
-Tests for 'All prefixes were successfully created. 200', 'Some prefix
-modifications failed. 207', '400: All modifications failed', and 'Unauthorized. Authentication credentials were
+Tests for 'All prefixes were successfully created. 200', 'Some or all prefix
+creations failed. 207', and 'Unauthorized. Authentication credentials were
 not provided. 401'
 
 For the 207 response Each object submitted will have it's own response object
@@ -29,45 +29,50 @@ class CreatePrefixeTestCase(APITestCase):
 
         self.client= APIClient()
         self.data = [{
-            "prefix": "test",
+            "prefix": "test1",
             "description": "Test prefix description.",
-            "authorized_groups": ["bco_publisher", "bco_drafter"]
+            "public": "true"
+        },
+        {
+            "prefix": "test2",
+            "description": "Test prefix description.",
+            "public": "true"
         }]
 
         self.legacy_data = {
-            "POST_api_prefixes_modify": [
+            "POST_api_prefixes_create": [
                 {
                     "owner_group": "bco_publisher",
                     "owner_user": "bco_api_user",
                     "prefixes": [
                         {
-                            "description": "Just a test modification for prefix.",
-                            "prefix": "Test"
+                            "description": "Just a test prefix.",
+                            "prefix": "testR"
                         }
                     ]
                 }
             ]
         }
 
-    def test_modify_prefix_success(self):
-        """The prefix was successfully modified. 200
+    def test_create_prefix_success(self):
+        """The prefix was successfully created. 200
         """
 
-        token = Token.objects.get(user=User.objects.get(username='tester')).key
+        token = Token.objects.get(user=User.objects.get(username='bco_api_user')).key
 
         self.client.credentials(HTTP_AUTHORIZATION='Token ' + token)
-        legacy_response = self.client.post('/api/prefixes/modify/', data=self.legacy_data, format='json')
-        response = self.client.post('/api/prefixes/modify/', data=self.data, format='json')
+        legacy_response = self.client.post('/api/prefixes/create/', data=self.legacy_data, format='json')
+        response = self.client.post('/api/prefixes/create/', data=self.data, format='json')
         self.assertEqual(legacy_response.status_code, 200)
         self.assertEqual(response.status_code, 200)
 
-    def test_modify_multi_status(self):
-        """Tests for 'Some prefix modifications failed. 207.'
+    def test_create_multi_status(self):
+        """Tests for 'Some prefix creations failed. 207.'
         """
 
-        token = Token.objects.get(user=User.objects.get(username='tester')).key        
+        token = Token.objects.get(user=User.objects.get(username='bco_api_user')).key        
         data = {
-            "POST_api_prefixes_modify": [
+            "POST_api_prefixes_create": [
                 {
                     "owner_group": "test_drafter",
                     "owner_user": "bco_api_user",
@@ -94,8 +99,8 @@ class CreatePrefixeTestCase(APITestCase):
                     "owner_user": "bco_api_user",
                     "prefixes": [
                         {
-                            "description": "Just a test prefix update.",
-                            "prefix": "test"
+                            "description": "Just a test prefix.",
+                            "prefix": "testR"
                         },
                         
                     ]
@@ -106,7 +111,7 @@ class CreatePrefixeTestCase(APITestCase):
                     "prefixes": [
                         {
                             "description": "Just a test prefix.",
-                            "prefix": "BCO"
+                            "prefix": "test"
                         }
                     ]
                 }
@@ -114,17 +119,15 @@ class CreatePrefixeTestCase(APITestCase):
         }
 
         self.client.credentials(HTTP_AUTHORIZATION='Token ' + token)
-        response = self.client.post('/api/prefixes/modify/', data=data, format='json')
+        response = self.client.post('/api/prefixes/create/', data=data, format='json')
         # 201: The prefix * was successfully created.
-        self.assertEqual(response.data[2]['TEST']['status_code'], 200)
+        self.assertEqual(response.data[2]['TESTR']['status_code'], 201)
 
-        # 400: Bad Request. The prefix * does not exist.
+        # 400: Bad Request. The prefix * does not follow the naming rules for a prefix.
         self.assertIn('prefix', response.data[0]['INVALID-PREFIX']['data'])
-        # 404: Not Found. The user * was not found on the server.
-        self.assertIn('authorized_groups', response.data[1]['TESTR']['data'])
         
         # 409: Conflict. The prefix the requestor is attempting to create already exists.
-        self.assertIn('permissions', response.data[3]['BCO']['message'])
+        self.assertIn('prefix_name', response.data[3]['TEST']['data'])
     
         self.assertEqual(response.status_code, 207)
